+++ pysui/sui/sui_clients/async_client.py
@@ -1,4 +1,4 @@
-#    Copyright 2022 Frank V. Castellucci
+#    Copyright Frank V. Castellucci
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
 #    You may obtain a copy of the License at
@@ -135,46 +135,13 @@ class SuiClient(_ClientMixin):
                 msg = result.result_data["error"]["message"]
                 if msg == _ClientMixin._SIGNATURE_ERROR and recovery_id > 0:
                     result = SuiRpcResult(False, msg)
-                elif msg == _ClientMixin._SIGNATURE_ERROR and recovery_id == 0:
-                    result = await _inner_sign(1)
+                elif msg == _ClientMixin._SIGNATURE_ERROR and recovery_id < 0:
+                    result = await _inner_sign(recovery_id + 1)
                 else:
                     result = SuiRpcResult(False, msg)
             return result
 
         return await _inner_sign(_recovery_id)
-
-    # async def sign_and_submit(self, signer: SuiAddress, tx_bytes: SuiTxBytes) -> Union[SuiRpcResult, Exception]:
-    #     """sign_and_submit Signs the transaction bytes from previous submission, signs and executes.
-
-    #     :param signer: Signer for transaction. Should be the same from original transaction
-    #     :type signer: SuiAddress
-    #     :param tx_bytes: Transaction bytes from previous submission
-    #     :type tx_bytes: SuiTxBytes
-    #     :return: Result from execution
-    #     :rtype: Union[SuiRpcResult, Exception]
-    #     """
-    #     kpair = self.config.keypair_for_address(signer)
-    #     if self.version_at_least(0, 22, 0):
-    #         builder = ExecuteTransaction(
-    #             tx_bytes=tx_bytes,
-    #             signature=kpair.new_sign_secure(tx_bytes.tx_bytes),
-    #             request_type=self.request_type,
-    #         )
-    #     elif self.version_at_least(0, 18, 0):
-    #         builder = ExecuteSerializedTransaction(
-    #             tx_bytes=tx_bytes,
-    #             signature=kpair.new_sign_secure(tx_bytes.tx_bytes),
-    #             request_type=self.request_type,
-    #         )
-    #     else:
-    #         return SuiRpcResult(False, "Unsupported SUI API version")
-    #     result = await self._execute(builder)
-    #     if result.is_ok():
-    #         if "error" in result.result_data:
-    #             return SuiRpcResult(False, result.result_data["error"]["message"], None)
-    #         # print(json.dumps(result.result_data["result"], indent=2))
-    #         result = SuiRpcResult(True, None, builder.handle_return(result.result_data["result"]))
-    #     return result
 
     async def dry_run(self, builder: SuiBaseBuilder) -> Union[SuiRpcResult, Exception]:
         """Submit transaction than sui_dryRunTransaction only."""